Changed clickable links on tweet
To allow for a better experience and make it more easy to click on tweets, all text views in the timeline do not allow user interaction. In order for the user to click on a link, he must first go to the tweet's detail page and click on it there.